Reverse order of imports
This way, we will import packages which are present in the virtual env first,
then injecteted packages into virtual environment.

This is a fix for an issue where we analyze click package, which is also a
dependency of thoth-solver. As the precedence is given first to path injected
by PYTHONPATH environment variable, we always return metadata for click which
is installed for thoth-solver (not the analyzed one).